Agent

R6 class Agent oversees the running of one Bandit and Policy pair over each time step t. A Simulator instance generally runs multiple Agents n parallel.

Usage

agent <- Agent$new(policy, bandit, name=NULL, sparse = 0.0)

Arguments

policy

Policy instance.

bandit

Bandit instance.

name

character; sets the name of the Agent. If NULL (default), Agent generates a name based on its Policy's name.

sparse

numeric; artificially reduces the data size by setting a sparsity level for the current Bandit and Policy pair. When set to a value between 0.0 (default) and 1.0 only a fraction sparse of the Bandit's data is randomly chosen to be available to improve the Agent's Policy through policy$set_reward.

Methods

new()

generates and instantializes a new Agent instance.

do_step()

advances a simulation by one time step by consecutively calling bandit$get_context(), policy$get_action(), bandit$get_reward() and policy$set_reward(). Returns a list of lists with context, action, reward and theta.

set_t(t)

integer; Sets the current time step to t.

get_t()

returns the current time step t.

See also

Examples

  policy    <- EpsilonGreedyPolicy$new(epsilon = 0.1)
  bandit    <- BasicBernoulliBandit$new(weights = c(0.6, 0.1, 0.1))

#> Error in .subset2(public_bind_env, "initialize")(...): unused argument (weights = c(0.6, 0.1, 0.1))

  agent     <- Agent$new(policy, bandit, name = "E.G.", sparse = 0.5)

#> Error in .subset2(public_bind_env, "initialize")(...): object 'bandit' not found

  history   <- Simulator$new(agents = agent,
                             horizon = 10,
                             simulations = 10,
                             progress_file = TRUE)$run()

#> Error in .subset2(public_bind_env, "initialize")(...): object 'agent' not found