Contextual Utility and Root Class

The R6 class Contextual is the root of the {contextual} package's class hierarchy. That is, every class in the {contextual} package has Contextual as a superclass. Contextual contains several utility functions that are used throughtout {contextual}.

Usage

contextual <- Contextual$new()

Package

The {contextual} package enables you to simulate and analyze contextual multi-armed bandit algorithms with poise.

Arguments

hash

Returns an unique hash for any instantiated {contextual} child object. This hash value is acquired from the R6 object's environment.

Methods

new()

Generates a new Contextual object.

max_in(x, equal_is_random = TRUE))

Returns the index of the maximum value in list x. If there is a tie and equal_is_random is TRUE, the index of one of the tied maxima is returned at random. If equal_is_random is FALSE, the maximum with the lowest index number is returned.

sum_of(x)

Returns the sum of the values of the elements of a list x.

inc(x)<-value

Increments x by value. Equivalent to x <- x + value. (orginal: Hmisc)

dec(x)<-value

Decrements x by value. Equivalent to x <- x - value. (orginal: Hmisc)

References

Winston Chang (2017). R6: Classes with Reference Semantics. R package version 2.2.2. https://CRAN.R-project.org/package=R6

Frank E Harrell Jr, with contributions from Charles Dupont and many others. (2017). Hmisc: Harrell Miscellaneous. R package version 4.0-3. https://CRAN.R-project.org/package=Hmisc

See also

Online: Documentation

Other contextual classes: AbstractBandit

Examples

contextual <- Contextual$new()

print(contextual$hash)

#> [1] "0x000000001a063ae8"

theta = list(par_one = list(1,2,3), par_two = list(2,3,4))

contextual$max_in(theta$par_one)

#> [1] 3

contextual$sum_of(theta$par_one)

#> [1] 6

# For this example section, the inc<- and dec<-
# functions need to be extracted from our
# contextual object.

"inc<-" <- contextual$`inc<-`
"dec<-" <- contextual$`dec<-`

# Within our contextual class hierarchy,
# these functions do not need to be
# extracted and can be applied directly
# as illustrated below.

x <- 5

inc(x) <- 1
x == 6

#> [1] TRUE

dec(x) <- 2
x == 4

#> [1] TRUE