Policy: A Time and Space Efficient Algorithm for Contextual Linear Bandits

Policy: A Time and Space Efficient Algorithm for Contextual Linear Bandits

Usage

 policy <- ContextualEpochGreedyPolicy$new(sZl = 10)

See also