Policy: Contextual Thompson Sampling with Linear Payoffs

ContextualThompsonSamplingPolicy works by maintaining a prior on the the mu_hat rewards of its arms. In this, it follows a beta–binomial model with parameters alpha and beta, sampling values for each arm from its prior and picking the arm with the highest value. When an arm is pulled and a Bernoulli reward is observed, it modifies the prior based on the reward. This procedure is repeated for the next arm pull.

Usage

policy <- ContextualThompsonSamplingPolicy(alpha = 1, beta = 1, name = "TSampling")

Arguments

alpha

integer, a natural number N>0 - first parameter of the Beta distribution

beta

integer, a natural number N>0 - second parameter of the Beta distribution

name

character string specifying this policy. name is, amongst others, saved to the History log and displayed in summaries and plots.

Methods

new(alpha = 1, beta = 1, name = "TSampling")

Generates a new ContextualThompsonSamplingPolicy object. Arguments are defined in the Argument section above.

set_parameters()

each policy needs to assign the parameters it wants to keep track of to list self$theta_to_arms that has to be defined in set_parameters()'s body. The parameters defined here can later be accessed by arm index in the following way: theta[[index_of_arm]]$parameter_name

get_action(context)

here, a policy decides which arm to choose, based on the current values of its parameters and, potentially, the current context.

set_reward(reward, context)

in set_reward(reward, context), a policy updates its parameter values based on the reward received, and, potentially, the current context.

References

Thompson, W. R. (1933). On the likelihood that one unknown probability exceeds another in view of the evidence of two samples. Biometrika, 25(3/4), 285-294.

See also

Core contextual classes: Contextual, Simulator, Agent, History, Plot

Bandit classes: AbstractBandit, BasicBandit, OfflineLiBandit, SyntheticBandit

Other contextual policies: AbstractPolicy, BTSLogitPolicy, ContextualEpsilonGreedyPolicy, EpsilonFirstPolicy, EpsilonGreedyPolicy, Exp3Policy, LinUCBPolicy, OraclePolicy, RandomPolicy, SimpleBTSPolicy, SoftmaxPolicy, ThompsonSamplingPolicy, UCB1Policy