History

The R6 class History keeps a log of all Simulator interactions in its internal data.table. It also provides basic data summaries, and can save or load simulation log data files.

Usage

History <- History$new(n = 1, save_context = FALSE, save_theta = FALSE)

Arguments

n

integer. The number of rows, to be preallocated during initialization.

save_context

logical. Save context vectors X when writing simulation data?

save_theta

logical. Save parameter lists theta when writing simulation data?

Methods

reset()

Resets a History instance to its original initialisation values.

save(index, t, action, reward, policy_name, simulation_index, context_value = NA, theta_value = NA)

Saves one row of simulation data.

save_data(filename = NA)

Writes the History log file in its default data.table format, with filename as the name of the file which the data is to be written to.

load_data = function(filename, nth_rows = 0)

Reads a History log file in its default data.table format, with filename as the name of the file which the data are to be read from. If nth_rows is larger than 0, every nth_rows of data is read instead of the full data file. This can be of use with (a first) analysis of very large data files.

get_data_frame()

Returns the History log as a data.frame.

get_data_table()

Returns the History log as a data.table.

set_data_table(dt)

Sets the History log with the data in data.table dt.

clear_data_table()

Clears the History log.

delete_empty_rows()

Deletes all empty rows in the History log and reindexes the t column grouped by agent and simulation.

reindex_t(truncate = TRUE)

Removes empty rows from the History log, reindexes the t column, and, if truncate is TRUE, truncates the resulting data to the shortest simulation grouped by agent and simulation.

print_data()

Prints a summary of the History log.

cumulative(final = TRUE, regret = TRUE, rate = FALSE)

Returns cumulative reward (when regret is FALSE) or regret. When final is TRUE, it only returns the final value. When final is FALSE, it returns a data.table containing all cumulative reward or regret values from 1 to T. When rate is TRUE, cumulative reward or regret are divided by column t before any values are returned.

See also